Friday/requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
nbformat
jupyter
Pillow
